refactor(workflows): group run UI under components/runs, one component per file
Move the run list/detail components into modules/workflows/components/runs and
split the detail drawer's inline sections into their own files, following the
one-component-per-file convention:

- runs/workflow-runs-table.tsx
- runs/workflow-run-detail-drawer.tsx
- runs/workflow-run-steps.tsx
- runs/run-summary-section.tsx
- runs/run-steps-body.tsx
- runs/run-json-section.tsx

Update the run page imports to the new paths. No behavior change.
